feat: Move all stubs to a dedicated dir
Part of #81
