+++ tests/unit/test_tools.h
@@ -21,11 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#include <gtest/gtest.h>
-#include <gmock/gmock.h>
-#include "MessageCollector.h"
-#include <sstream>
 #include "AST.h"
+#include "MessageCollector.h"
+#include "stubs.h"
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
+#include <sstream>
 
 using namespace ::testing;
 
@@ -35,9 +36,10 @@ using namespace ::testing;
 
 #define DEFINE_ENVIRONMENT(name) auto *name = new filc::environment::Environment("name", filc::environment::Environment::getGlobalEnvironment())
 
-#define DEFINE_LLVM auto *context = new llvm::LLVMContext; \
-                    auto *module = new llvm::Module("module", *context); \
-                    auto *builder = new llvm::IRBuilder<>(*context)
+#define DEFINE_LLVM                                       \
+    auto *context = new llvm::LLVMContext;                \
+    auto *module  = new llvm::Module("module", *context); \
+    auto *builder = new llvm::IRBuilder<>(*context)
 
 #define ASSERT_TYPE(expected, type) ASSERT_STREQ(expected, type->dump().c_str())
 
@@ -49,55 +51,33 @@ using namespace ::testing;
 
 #define ASSERT_LITERAL(expected, type, literal) ASSERT_EQ(expected, dynamic_cast<filc::ast::type *>(literal)->getValue())
 
-#define ASSERT_VARIABLE_DECLARATION(constant, name, type, value, literal, variable) \
-    auto *var_##variable = dynamic_cast<filc::ast::VariableDeclaration *>(variable);\
-    ASSERT_NE(nullptr, var_##variable); \
-    ASSERT_EQ(constant, var_##variable->isConstant()); \
-    ASSERT_IDENTIFIER(name, var_##variable->getIdentifier()); \
-    ASSERT_TYPE(type, var_##variable->getType()); \
+#define ASSERT_VARIABLE_DECLARATION(constant, name, type, value, literal, variable)  \
+    auto *var_##variable = dynamic_cast<filc::ast::VariableDeclaration *>(variable); \
+    ASSERT_NE(nullptr, var_##variable);                                              \
+    ASSERT_EQ(constant, var_##variable->isConstant());                               \
+    ASSERT_IDENTIFIER(name, var_##variable->getIdentifier());                        \
+    ASSERT_TYPE(type, var_##variable->getType());                                    \
     ASSERT_LITERAL(value, literal, var_##variable->getAssignation())
 
-#define ASSERT_MESSAGE_CONTENT(expected, message) { \
-    std::stringstream stream; \
-    message.print(stream); \
-    std::string result(std::istreambuf_iterator<char>(stream), {}); \
-    ASSERT_STREQ(expected, result.c_str()); \
+#define ASSERT_MESSAGE_CONTENT(expected, message)                       \
+    {                                                                   \
+        std::stringstream stream;                                       \
+        message.print(stream);                                          \
+        std::string result(std::istreambuf_iterator<char>(stream), {}); \
+        ASSERT_STREQ(expected, result.c_str());                         \
     }
 
-#define ASSERT_OUTPUT(expected, expression) { \
-    std::stringstream redirect_stream; \
-    std::streambuf* oldbuf = std::cout.rdbuf(redirect_stream.rdbuf()); \
-    expression; \
-    std::string result, line; \
-    while (std::getline(redirect_stream, line)) { \
-        result += line + "\n"; \
-    } \
-    ASSERT_STREQ(expected, result.c_str()); \
-    std::cout.rdbuf(oldbuf); \
+#define ASSERT_OUTPUT(expected, expression)                                \
+    {                                                                      \
+        std::stringstream redirect_stream;                                 \
+        std::streambuf *oldbuf = std::cout.rdbuf(redirect_stream.rdbuf()); \
+        expression;                                                        \
+        std::string result, line;                                          \
+        while (std::getline(redirect_stream, line)) {                      \
+            result += line + "\n";                                         \
+        }                                                                  \
+        ASSERT_STREQ(expected, result.c_str());                            \
+        std::cout.rdbuf(oldbuf);                                           \
     }
 
 auto redirectCin(const std::function<void(std::stringstream &stream)> &function) -> void;
-
-class TestExpression : public filc::ast::AbstractExpression {
-public:
-    TestExpression();
-
-    auto withExpressionType(const std::shared_ptr<filc::ast::AbstractType> &expression_type) -> TestExpression &;
-
-    auto resolveType(filc::environment::Environment *environment,
-                     filc::message::MessageCollector *collector,
-                     const std::shared_ptr<filc::ast::AbstractType> &preferred_type) -> void override;
-
-    [[nodiscard]] auto isResolveTypeCalled() const -> bool;
-
-    auto addNameToEnvironment(filc::environment::Environment *environment) const -> void override;
-
-    auto generateIR(filc::message::MessageCollector *collector,
-                    filc::environment::Environment *environment,
-                    llvm::LLVMContext *context,
-                    llvm::Module *module,
-                    llvm::IRBuilder<> *builder) const -> llvm::Value * override;
-
-private:
-    bool _resolveType_called;
-};
