+++ tests/unit/utils/command/HelpCommandTest.cpp
@@ -22,7 +22,6 @@
  * SOFTWARE.
  */
 #include "HelpCommand.h"
-#include "CustomCommand.h"
 #include "test_tools.h"
 
 using namespace filc::utils::command;
